requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
aiohttp>=3.9.1
langchain>=0.0.350
transformers>=4.36.2
huggingface-hub>=0.20.3
torch>=2.1.2
pydantic>=2.5.2
coinbase
psutil>=5.9.7
together